tidy up some more
