+++ client/engine/chainservice/adjudicator/doc.go
@@ -1,0 +1,2 @@
+// package NitroAdjudicator contains bindings generated from the Nitro Adjudicator solidity contract, as well as utilities for converting client types to on-chain types.
+package NitroAdjudicator
